Start adding account and entry management.

+++ database/migrations/2021_01_14_194149_create_entry_credits_table.php
@@ -15,7 +15,7 @@ class CreateEntryCreditsTable extends Migration
     {
         Schema::create('ledger_entry_credits', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('amount');
+            $table->unsignedBigInteger('amount');
             $table->unsignedBigInteger('entry_id');
             $table->unsignedBigInteger('account_id');
             $table->timestamps();
